Refactor the test directories
