+++ watson-assistant-cfenv-processor/src/test/java/com/ibm/boot/cfenvprocessors/WatsonAssistantCfEnvProcessorApplicationTests.java
@@ -1,0 +1,14 @@
+package com.ibm.boot.cfenvprocessors;
+
+import org.junit.Test;
+import org.junit.runner.RunWith;
+import org.springframework.test.context.junit4.SpringRunner;
+
+@RunWith(SpringRunner.class)
+public class WatsonAssistantCfEnvProcessorApplicationTests {
+
+    @Test
+    public void contextLoads() {
+    }
+
+}
