+++ watson-compare-comply-cfenv-processor/src/test/java/com/ibm/boot/cfenvprocessors/WatsonCompareComplyCfEnvProcessorApplicationTests.java
@@ -1,0 +1,14 @@
+package com.ibm.boot.cfenvprocessors;
+
+import org.junit.Test;
+import org.junit.runner.RunWith;
+import org.springframework.test.context.junit4.SpringRunner;
+
+@RunWith(SpringRunner.class)
+public class WatsonCompareComplyCfEnvProcessorApplicationTests {
+
+    @Test
+    public void contextLoads() {
+    }
+
+}
