--- watson-compare-comply-cfenv-processor/src/test/java/org/terrence/cfenvprocessors/WatsonCompareComplyCfEnvProcessorApplicationTests.java
@@ -1,14 +1,0 @@
-package org.terrence.cfenvprocessors;
-
-import org.junit.Test;
-import org.junit.runner.RunWith;
-import org.springframework.test.context.junit4.SpringRunner;
-
-@RunWith(SpringRunner.class)
-public class WatsonCompareComplyCfEnvProcessorApplicationTests {
-
-    @Test
-    public void contextLoads() {
-    }
-
-}
