+++ watson-natural-language-classifier-cfenv-processor/src/test/java/com/ibm/boot/cfenvprocessors/WatsonNaturalLanguageClassifierCfEnvProcessorApplicationTests.java
@@ -1,0 +1,14 @@
+package com.ibm.boot.cfenvprocessors;
+
+import org.junit.Test;
+import org.junit.runner.RunWith;
+import org.springframework.test.context.junit4.SpringRunner;
+
+@RunWith(SpringRunner.class)
+public class WatsonNaturalLanguageClassifierCfEnvProcessorApplicationTests {
+
+    @Test
+    public void contextLoads() {
+    }
+
+}
