--- watson-natural-language-understanding-cfenv-processor/src/test/java/org/terrence/cfenvprocessors/WatsonNaturalLanguageUnderstandingCfEnvProcessorApplicationTests.java
@@ -1,14 +1,0 @@
-package org.terrence.cfenvprocessors;
-
-import org.junit.Test;
-import org.junit.runner.RunWith;
-import org.springframework.test.context.junit4.SpringRunner;
-
-@RunWith(SpringRunner.class)
-public class WatsonNaturalLanguageUnderstandingCfEnvProcessorApplicationTests {
-
-    @Test
-    public void contextLoads() {
-    }
-
-}
